translate error messages of shogi package
