+++ src/common/i18n/language.ts
@@ -1,0 +1,5 @@
+export enum Language {
+  // ISO 639-1
+  JA = "ja",
+  EN = "en",
+}
